+++ fj-daogen-sample/src/main/java/org/fugerit/java/daogen/sample/def/facade/EntityUserFacade.java
@@ -1,21 +1,11 @@
 package org.fugerit.java.daogen.sample.def.facade;
 
-// custom import start ( code above here will be overwritten )
-// custom import end ( code below here will be overwritten )
-
 /**
  * EntityUserFacade, version : 1.0.0
  *
  * author: fugerit
- *
- * warning!: auto generated object, insert custom code only between comments :
- * // custom code start ( code above here will be overwritten )
- * // custom code end ( code below here will be overwritten )
  */
 public interface EntityUserFacade extends EntityUserFacadeHelper {
-
-	// custom code start ( code above here will be overwritten )
-	// custom code end ( code below here will be overwritten )
 
 	// [HELPER/IMPL MODEL] this class is a stub and can be modified as you see fit (it will not been overwritten)
 }