+++ fj-daogen-sample/src/main/java/org/fugerit/java/daogen/sample/impl/facade/data/DataEntityLogDataFacade.java
@@ -2,24 +2,14 @@ package org.fugerit.java.daogen.sample.impl.facade.data;
 
 import org.fugerit.java.daogen.sample.def.facade.EntityLogDataFacade;
 
-// custom import start ( code above here will be overwritten )
-// custom import end ( code below here will be overwritten )
-
 /**
  * DataEntityLogDataFacade, version : 1.0.0
  *
  * author: fugerit
- *
- * warning!: auto generated object, insert custom code only between comments :
- * // custom code start ( code above here will be overwritten )
- * // custom code end ( code below here will be overwritten )
  */
 public class DataEntityLogDataFacade extends DataEntityLogDataFacadeHelper implements EntityLogDataFacade {
 
-	// custom code start ( code above here will be overwritten )
-	// custom code end ( code below here will be overwritten )
-
-	private static final long serialVersionUID = 84221996652L;
+	private static final long serialVersionUID = 806676938185L;
 
 	// [HELPER/IMPL MODEL] this class is a stub and can be modified as you see fit (it will not been overwritten)
 }