+++ fj-daogen-sample/src/main/java/org/fugerit/java/daogen/sample/impl/facade/data/FugeritDataLogicFacade.java
@@ -1,26 +1,13 @@
 package org.fugerit.java.daogen.sample.impl.facade.data;
 
-// custom import start ( code above here will be overwritten )
-// custom import end ( code below here will be overwritten )
-
 /**
  * FugeritDataLogicFacade, version : 1.0.0
  *
  * author: fugerit
- *
- * warning!: auto generated object, insert custom code only between comments :
- * // custom code start ( code above here will be overwritten )
- * // custom code end ( code below here will be overwritten )
  */
-public class FugeritDataLogicFacade extends FugeritDataLogicFacadeHelper {
+public class FugeritDataLogicFacade extends FugeritDataLogicFacadeHelper implements org.fugerit.java.daogen.sample.def.facade.FugeritLogicFacade, java.io.Serializable {
 
-	/**
-	 * 
-	 */
-	private static final long serialVersionUID = -3428899643186915522L;
-
-	// custom code start ( code above here will be overwritten )
-	// custom code end ( code below here will be overwritten )
+	private static final long serialVersionUID = 916703990459L;
 
 	// [HELPER/IMPL MODEL] this class is a stub and can be modified as you see fit (it will not been overwritten)
 }