+++ fj-daogen-sample/src/main/java/org/fugerit/java/daogen/sample/impl/rest/load/LoadTestTwoFieldKey.java
@@ -20,7 +20,7 @@ public class LoadTestTwoFieldKey extends LoadTestTwoFieldKeyHelper {
 	// custom code start ( code above here will be overwritten )
 	// custom code end ( code below here will be overwritten )
 
-	private static final long serialVersionUID = 892898388143L;
+	private static final long serialVersionUID = 351367114180L;
 
 	// [HELPER/IMPL MODEL] this class is a stub and can be modified as you see fit (it will not been overwritten)
 }