+++ fj-daogen-sample/src/main/java/org/fugerit/java/daogen/sample/impl/rest/load/LoadUser.java
@@ -20,7 +20,7 @@ public class LoadUser extends LoadUserHelper {
 	// custom code start ( code above here will be overwritten )
 	// custom code end ( code below here will be overwritten )
 
-	private static final long serialVersionUID = 609752129833L;
+	private static final long serialVersionUID = 785780699289L;
 
 	// [HELPER/IMPL MODEL] this class is a stub and can be modified as you see fit (it will not been overwritten)
 }