+++ fj-daogen-sample/src/main/java/org/fugerit/java/daogen/sample/impl/rest/load/LoadUserData.java
@@ -20,7 +20,7 @@ public class LoadUserData extends LoadUserDataHelper {
 	// custom code start ( code above here will be overwritten )
 	// custom code end ( code below here will be overwritten )
 
-	private static final long serialVersionUID = 187757514010L;
+	private static final long serialVersionUID = 841353578465L;
 
 	// [HELPER/IMPL MODEL] this class is a stub and can be modified as you see fit (it will not been overwritten)
 }