+++ fj-daogen-sample/src/main/java/org/fugerit/java/daogen/sample/impl/struct/ObjAddress.java
@@ -44,7 +44,7 @@ public class ObjAddress extends WrapperAddress implements SQLData, StructMapper 
 
 	@Override
 	public Map<String, Class<?>> newTypeMapper() throws SQLException {
-		Map<String, Class<?>> map = new HashMap<String, Class<?>>();
+		Map<String, Class<?>> map = new HashMap<>();
 		map.put( SQL_TYPE_NAME, ObjAddress.class );
 		return map;
 	}