+++ fj-daogen-sample/src/main/java/org/fugerit/java/daogen/sample/impl/struct/ObjUser.java
@@ -44,7 +44,7 @@ public class ObjUser extends WrapperUser implements SQLData, StructMapper {
 
 	@Override
 	public Map<String, Class<?>> newTypeMapper() throws SQLException {
-		Map<String, Class<?>> map = new HashMap<String, Class<?>>();
+		Map<String, Class<?>> map = new HashMap<>();
 		map.put( SQL_TYPE_NAME, ObjUser.class );
 		return map;
 	}